fix: 전수 감사 보안/동시성/성능 결함 일괄 수정 (~163건)

+++ apps/backend/src/main/java/world/yeon/backend/googledrive_browser/dto/GoogleDriveFileResponse.java
@@ -1,3 +1,3 @@
 package world.yeon.backend.googledrive_browser.dto;
 
-public record GoogleDriveFileResponse(String id, String name, int size, String lastModifiedAt, String mimeType) {}
+public record GoogleDriveFileResponse(String id, String name, long size, String lastModifiedAt, String mimeType) {}
